+++ ocaml-lsp-server/src/ocaml_lsp_server.ml
@@ -103,6 +103,8 @@ let initialize_info (client_capabilities : ClientCapabilities.t) : InitializeRes
               ; Req_type_expression.capability
               ; Req_locate.capability
               ; Req_destruct.capability
+              ; Req_locate_types.capability
+              ; Req_refactor_extract.capability
               ] )
         ]
     in
@@ -151,6 +153,7 @@ let initialize_info (client_capabilities : ClientCapabilities.t) : InitializeRes
       ~referencesProvider:(`Bool true)
       ~documentHighlightProvider:(`Bool true)
       ~documentFormattingProvider:(`Bool true)
+      ~documentRangeFormattingProvider:(`Bool true)
       ~selectionRangeProvider:(`Bool true)
       ~documentSymbolProvider:(`Bool true)
       ~workspaceSymbolProvider:(`Bool true)
@@ -365,6 +368,26 @@ module Formatter = struct
          let+ to_ = Dune.Instance.format_dune_file dune doc in
          Some (Diff.edit ~from:(Document.text doc) ~to_))
   ;;
+
+  let run_on_range rpc doc range =
+    let* res =
+      let* cancel = Server.cancel_token () in
+      Ocamlformat.run_on_range doc range cancel
+    in
+    match res with
+    | Ok result -> Fiber.return (Some result)
+    | Error e ->
+      let+ () =
+        let state : State.t = Server.state rpc in
+        let msg =
+          let message = Ocamlformat.message e in
+          ShowMessageParams.create ~message ~type_:Warning
+        in
+        task_if_running state.detached ~f:(fun () ->
+          Server.notification rpc (ShowMessage msg))
+      in
+      Jsonrpc.Response.Error.raise (jsonrpc_error e)
+  ;;
 end
 
 let text_document_lens
@@ -425,7 +448,7 @@ let selection_range
           Document.Merlin.dispatch_exn
             ~name:"shape"
             merlin
-            (Enclosing (Position.logical x))
+            (Enclosing (Position.logical x, None))
         in
         selection_range_of_enclosings enclosings)
     in
@@ -562,6 +585,8 @@ let on_request
          ; Req_type_expression.meth, Req_type_expression.on_request
          ; Req_locate.meth, Req_locate.on_request
          ; Req_destruct.meth, Req_destruct.on_request
+         ; Req_locate_types.meth, Req_locate_types.on_request
+         ; Req_refactor_extract.meth, Req_refactor_extract.on_request
          ]
          meth
      with
@@ -715,6 +740,12 @@ let on_request
          let doc = Document_store.get store uri in
          Formatter.run rpc doc)
       ()
+  | TextDocumentRangeFormatting { textDocument = { uri }; range; _ } ->
+    later
+      (fun _ () ->
+         let doc = Document_store.get store uri in
+         Formatter.run_on_range rpc doc range)
+      ()
   | TextDocumentOnTypeFormatting _ -> now None
   | SelectionRange req -> later selection_range req
   | TextDocumentImplementation _ -> not_supported ()
@@ -722,7 +753,6 @@ let on_request
   | SemanticTokensDelta p -> later Semantic_highlighting.on_request_full_delta p
   | TextDocumentMoniker _ -> not_supported ()
   | TextDocumentPrepareCallHierarchy _ -> not_supported ()
-  | TextDocumentRangeFormatting _ -> not_supported ()
   | CallHierarchyIncomingCalls _ -> not_supported ()
   | CallHierarchyOutgoingCalls _ -> not_supported ()
   | SemanticTokensRange _ -> not_supported ()